Update comment create service
