+++ app/lib/comments/create.rb
@@ -1,6 +1,19 @@
 module Comments
-  class Create < Base
+  class Create
     attr_reader :post, :is_private, :content, :commenter
+
+    def self.run!(post:, is_private:, content:, commenter:)
+      service = new(post: post, is_private: is_private, content: content, commenter: commenter)
+      service.validate!
+      service.run!
+    end
+
+    def initialize(post:, is_private:, content:, commenter:)
+      @post = post
+      @is_private = is_private || false
+      @content = content
+      @commenter = commenter || Current.user
+    end
 
     def validate!
       validate_user_has_permission
@@ -42,6 +55,11 @@ module Comments
         visibility: @is_private ? Constants::Visibility::PRIVATE : Constants::Visibility::PUBLIC,
         post_id: @post.id
       )
+    end
+
+    def validate_user_has_permission
+      return if @is_private && Current.user.admin_of?(Current.company)
+      raise Errors::AdminLacksPermission
     end
   end
 end
